remove renaming transformer & keep default one for now
