+++ Quasar.Obfuscator/Quasar.Obfuscator.csproj
@@ -55,18 +55,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Core\Obfuscator.cs" />
-    <Compile Include="Core\Transformers\ITransformer.cs" />
-    <Compile Include="Core\Transformers\RenamerTransformer.cs" />
-    <Compile Include="Core\Transformers\StringEncryptionTransformer.cs" />
-    <Compile Include="Core\Utils\InjectHelper.cs" />
-    <Compile Include="Core\Utils\Injection\StringEncryption.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Core\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
